model의 cart 삭제 -> cartUtils로 이동
